Further migration of test code from legacy setters
This continues the migration of test code from legacy setters
and cosntructors to explicitly-typed ones.

+++ pcep/pcc-mock/src/test/java/org/opendaylight/protocol/pcep/pcc/mock/PCCSessionListenerTest.java
@@ -155,7 +155,7 @@ public class PCCSessionListenerTest {
             .setDelegate(true)
             .setIgnore(false)
             .setOperational(OperationalStatus.Up)
-            .setPlspId(new PlspId(999L))
+            .setPlspId(new PlspId(Uint32.valueOf(999)))
             .setProcessingRule(false)
             .setRemove(remove)
             .setSync(true);
